leverages a widget registry for custom SchemaField widgets
