+++ src/components/fields/schemaFields/CssSpacingField.tsx
@@ -15,7 +15,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-import { defaultFieldFactory } from "@/components/fields/schemaFields/SchemaFieldContext";
+import { defaultFieldFactory } from "./defaultFieldFactory";
 import CssSpacingWidget from "@/components/fields/schemaFields/widgets/cssClassWidgets/CssSpacingWidget";
 
 const CssSpacingField = defaultFieldFactory(CssSpacingWidget);
