+++ src/components/fields/schemaFields/HeadingStyleField.tsx
@@ -15,7 +15,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-import { defaultFieldFactory } from "@/components/fields/schemaFields/SchemaFieldContext";
+import { defaultFieldFactory } from "./defaultFieldFactory";
 import HeadingStyleWidget from "./widgets/HeadingStyleWidget";
 
 const HeadingStyleField = defaultFieldFactory(HeadingStyleWidget);
