+++ src/pageEditor/fields/devtoolFieldOverrides.tsx
@@ -23,6 +23,7 @@ import { type Schema } from "@/types/schemaTypes";
 import OptionIcon from "@/components/fields/schemaFields/optionIcon/OptionIcon";
 import { type CustomFieldDefinitions } from "@/components/fields/schemaFields/schemaFieldTypes";
 import { isTemplateExpression } from "@/utils/expressionUtils";
+import { customWidgets } from "@/components/fields/schemaFields/SchemaFieldContext";
 
 const ClearableSelectorWidget: React.FunctionComponent<
   SelectorSelectorProps
@@ -54,6 +55,7 @@ const devtoolFieldOverrides: CustomFieldDefinitions = {
       },
     },
   ],
+  customWidgets,
 };
 
 export default devtoolFieldOverrides;
